+++ frontend/src/App.jsx
@@ -1,13 +1,14 @@
 import { Routes, Route } from 'react-router-dom';
 import HomePage from './pages/HomePage';
+import ResultPage from './pages/ResultPage';
+
 import './App.css';
 
 export default function App() {
   return (
     <Routes>
       <Route path="/" element={<HomePage />} />
-      
-      {/*<Route path="/acesso/:chave" element={<aDefinir />} />*/}
+      <Route path="/acesso/:chave" element={<ResultPage />} />
     </Routes>
   );
 }
