+++ frontend/src/pages/HomePage.jsx
@@ -1,12 +1,21 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom"; // Importe o hook de navegação
+import { useNavigate } from "react-router-dom"; 
 
 import AboutUs from "../components/home/AboutUs";
-import HeroSection from "../components/home/HeroSection";
 import ApiKeySection from "../components/home/ApiKeySection";
+import DescriptionSection from "../components/home/DescriptionSection";
 
 export default function HomePage() {
   const [scrollY, setScrollY] = useState(0);
+  const navigate = useNavigate(); 
+
+  const handleReportGenerated = (reportData) => {
+    if (reportData && reportData.account_id) {
+      navigate(`/acesso/${reportData.account_id}`, { state: { reportData } });
+    } else {
+      alert("Não foi possível obter os dados do relatório para navegar.");
+    }
+  };
 
   const handleScroll = () => {
     setScrollY(window.scrollY);
@@ -29,10 +38,13 @@ export default function HomePage() {
 
   return (
     <div className="bg-black text-white font-sans" style={{ height: "220vh" }}>
-      <HeroSection textTranslateY={textTranslateY} textOpacity={textOpacity} />
-
+      <ApiKeySection 
+        textTranslateY={textTranslateY} 
+        textOpacity={textOpacity} 
+        onReportGenerated={handleReportGenerated}
+      />
       <div className="relative">
-        <ApiKeySection />
+        <DescriptionSection />
 
         <div className="absolute top-[180vh] w-full">
           <AboutUs />
